pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yawgbot"
version = "0.0.4"
authors = [
    { name = "Roberto Castellotti", email = "me@rcastellotti.dev" },
]
description = "yawgbot - yet another wg bot"
readme = "README.md"
license = { file = "LICENSE.md" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "beautifulsoup4==4.11.1",
    "certifi==2022.6.15",
    "charset-normalizer==2.1.1",
    "click==8.1.3",
    "Flask==2.2.2",
    "greenlet==1.1.3",
    "idna==3.3",
    "itsdangerous==2.1.2",
    "Jinja2==3.1.2",
    "MarkupSafe==2.1.1",
    "platformdirs==2.5.2",
    "python-dotenv==0.21.0",
    "PyYAML==6.0",
    "requests==2.28.1",
    "soupsieve==2.3.2.post1",
    "SQLAlchemy==1.4.41",
    "SQLAlchemy-Utils==0.38.3",
    "urllib3==1.26.12",
    "Werkzeug==2.2.2"
]
[project.scripts]
yawgbot-web = "yawgbot.app:run"
[project.urls]
"Homepage" = "https://github.com/rcastellotti/yawgbot"
"Bug Tracker" = "https://github.com/rcastellotti/yawgbot/issues"